refactor chat modules
